+++ py/src/braintrust/integrations/auto_test_scripts/test_auto_temporal.py
@@ -1,0 +1,16 @@
+"""Test auto_instrument for Temporal."""
+
+from braintrust.auto import auto_instrument
+from braintrust.integrations.temporal import BraintrustPlugin, setup_temporal
+
+
+results = auto_instrument()
+assert results.get("temporal") == True
+
+results2 = auto_instrument()
+assert results2.get("temporal") == True
+
+assert setup_temporal() == True
+assert BraintrustPlugin is not None
+
+print("SUCCESS")
